Solution to Calculate and Print the Average of Three Numbers

+++ Calculate and Print the Average of Three Numbers@@6788d5012db25efec946f944/code.c
@@ -1,8 +1,8 @@
 #include<stdio.h>
 int main(){
-    int a,b,c,d;
-    scanf("%d",&a,&b,&c);
+    float a,b,c,d;
+    scanf("%f",&a,&b,&c);
     d=a+b+c/3;
-    printf("Average: %d\n", d);
+    printf("Average: %.2f\n", d);
     return 0;
 }
